+++ Source/Drawing2D/Point.h
@@ -13,14 +13,14 @@ namespace XC
         {
         public:
             Point();
-            Point(int x, int y);
+            Point(double x, double y);
             ~Point();
 
         public:
-            const int & GetX() const;
-            void SetX(const int & value);
-            const int & GetY() const;
-            void SetY(const int & value);
+            const double & GetX() const;
+            void SetX(const double & value);
+            const double & GetY() const;
+            void SetY(const double & value);
 
             operator Size() const;
